boulot en cours sur le score

+++ feelhub-application/src/main/java/com/feelhub/domain/topic/TopicSentimentScoreCalculator.java
@@ -2,6 +2,8 @@ package com.feelhub.domain.topic;
 
 
 import com.feelhub.domain.feeling.Sentiment;
+import com.google.common.collect.Lists;
+import com.google.common.collect.Ordering;
 import org.joda.time.DateTime;
 
 import java.util.List;
@@ -13,9 +15,11 @@ public final class TopicSentimentScoreCalculator {
             return 0;
         }
         int score = 0;
+        //Sentiment lastSentiment = findLastSentiment(sentiments);
         for(Sentiment sentiment : sentiments) {
             score += sentiment.getSentimentValue().numericValue();
         }
-        return score / sentiments.size() * 100;
+        return (int) (((double) score / (double) sentiments.size()) * 100);
     }
+
 }
